Rename plugin_util with underscore, update references
